Working p2p poller implementation.
Used REQ<<->ROUTER, not sure if such is optimal, as now we have to
register with the poller each additional REQ as it is created. Using a
ROUTER<->ROUTER setup didn't work due to a flaw with ROUTER requiring
the server (thing being connect()ed to) to set an identifier. Looks like
the server then could forge the identifier causing conflicts and such.