Probably mostly finish wildcard except for the fact there is no key.
Since they searched with an incomplete key, being the point of this
feature, we don't have the whole key in order to generate the id which
is what we need to search by. We also wouldn't be able to decrypt the
data nor verify it.  Further features will have to be implemented to
deal with these issues.