Change upload process to continue uploading even if one block fails.
Optimal operation is that this is because it is already uploaded to the
network and no more want it. We don't want to trust nodes saying they
have it already to limit us from trying to store it, so I'm not sure
what to do to ensure upload. It will have to come with the trust model.