Switch updateable keys with a path to have epubkey stored.
This makes the encrypted public key stored with the updateable key,
regardless of path. This makes non empty paths work now without having
to fetch the null path entry first. I think this might be better trade
off (space usage for efficiency of network?). It is easier to code for
now at least.