Implement the start of some network data healing code.
This is not optimal in that it happens outside the core tunnel code,
which means that it is doing a whole other FindNode process when it
could have reused the one which fetched. To be able to do so though the
protocol will have to be changed a bit, and it would probably be best to
rewrite chord_tasks.py first because it is already overly complicated
from its original simple design that is ancient.