Add a timeout to open_channel for immediate PeerS in send_find_node(..).
This was to fix that with Tor at least as tested through proxychains,
Tor is not notifying us that the Peer is no longer reachable with
standard TCP protocol of BrokenPipe, and thus our Node would keep trying
to use that Peer, causing useless infinite open_channel requests to be
sent and as well not replacing that Peer with a new connection, thus
eventually having all the PeerS in our connection table dead. This will
might save on leaked memory/channels with misbehaving nodes in normal
mode as well, so I have not made this code --tormode dependent.