Fix up client_engine to be always available (instead delay request).
Now the client_engine will always be set when later some process wants
it. To do so, instead, the do_GET and do_POST are blocked until it is
there. It is necessary to do that blocking because Maalstroom is started
before the Node so that the browser can be opened right away and the
Maalstroom socket is already bound.