Enhance Content-Security-Policy to include actual url for plugin.
This fixes the browser plugin to work with content that is from a
redirect. This is needed until I figure out how to make the browser not
rewrite the URL back to non morphis:// url after a redirect. The
redirect is sending Location: morphis://..., but firefox replaces that
with http://localhost:4251/... in the url bar and as far as it cares for
Content-Security-Policy. This commit should be rolled back once the
browser plugin is enhanced to fix firefox messing up after redirect.