Fix shell.py to handle multiple characters per incoming packet.
Before it made the (wrong) assuption that only one character would be
sent per packet; if the user jams the keys fast enough, that is not
true. Now we build up a buffer and process it more as a stream. It is
still a bit messy as to save time I did not completely rewrite it. For
instance, it is possibly broken still in that it will not be able to
process characters in a packet after an enter character. I was not able
to get such a thing to happen with openssh.